==> 20_markergene.Rmd <==
---
title: "RNA-seq find marker gene"
output: html_document
date: '2022-06-20'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Create scRNA-seq UMAP
```{r}
# Load the directory #-----#-----#-----#
setwd(your_working_directory)          #Where you saved your data 
#**************************************#
#
#
# Load all libraries #-----#-----#-----#
library(Seurat); library(Matrix); library(tidyverse)
#***************************************#
```

```{r}
integration = readRDS("Integration_062022.rds") # The last file is created in the 10_integration step.
DimPlot(integration, group.by = "celltype", label = TRUE, repel = TRUE)

# Change clusters' names
Idents(integration) = integration@meta.data$celltype
allMG_integration = FindAllMarkers(integration, min.pct = 0.1, logfc.threshold = 0.25)
top200MG_integration = allMG_integration %>% group_by(cluster) %>% top_n(n=200, wt = avg_log2FC)

# Save data for downstream analysis
saveRDS(top200MG_integration[top200MG_integration$cluster == "Cortex",]$gene, file = "integration_Cortex_200.rds")
saveRDS(top200MG_integration[top200MG_integration$cluster == "Endodermis",]$gene, file = "integration_Endodermis_200.rds")
```
